requirements.txt
playwright
playwright-stealth
edge-tts
moviepy
beautifulsoup4
python-dotenv
google-generativeai
requests
schedule
streamlit
pandas